Fix codefactor issues
